Refactor settings with tab navigation and fix invites
